Excel and Methods, Tried Selenium
Changed CSV output to Excel. Should be able to output to multiple sheets in the same file.
Moved the bulk of the middle of the method to their own functions.
Tried to automate completely with selenium and just input the site with all the tournaments listed but it did not work. The dropdown was too difficult to navagate. Now speakscrape_one can do all in the format (url,tournament name)